+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/types.ts
@@ -1,13 +1,23 @@
+import { IButtonProps, IPivotItemProps } from '@fluentui/react'
 import { IColumn } from '@fluentui/react/lib/DetailsList'
-import { IBaseComponentProps } from '../types'
 import { ProjectListModel } from 'models'
+import { IBaseComponentProps } from '../types'
+
+export interface IProjectListView extends IPivotItemProps {
+  searchBoxPlaceholder?: string
+  filter?: (project?: ProjectListModel) => boolean
+  getHeaderButtonProps?: (
+    state: IProjectListState
+  ) =>
+    | IButtonProps
+    | {
+        [key: string]: string | number | boolean
+      }
+}
+
+export type ProjectListRenderMode = 'tiles' | 'list'
 
 export interface IProjectListProps extends IBaseComponentProps {
-  /**
-   * Loading text
-   */
-  loadingText: string
-
   /**
    * Sort by property
    */
@@ -42,6 +52,16 @@ export interface IProjectListProps extends IBaseComponentProps {
    * Columns
    */
   columns?: IColumn[]
+
+  /**
+   * Default view
+   */
+  defaultView?: string
+
+  /**
+   * Hide views
+   */
+  hideViews?: string[]
 }
 
 export interface IProjectListState {
@@ -73,12 +93,11 @@ export interface IProjectListState {
   /**
    * How the projects should be rendered. `tiles` or `list`
    */
-  renderAs?: 'tiles' | 'list'
-
+  renderAs?: ProjectListRenderMode
   /**
    * Current selected view
    */
-  selectedView?: string
+  selectedView?: IProjectListView
 
   /**
    * Is the current user in the PortfolioManagerGroup?